nightshade-tracker/requirements.txt
# Core dependencies
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0

# Deep learning (for adversarial perturbations)
torch>=2.0.0
torchvision>=0.15.0

# Image hashing
imagehash>=4.3.1

# Error correction codes
reedsolo>=1.7.0

# Database
sqlite3  # Built-in, but listed for clarity

# Utilities
tqdm>=4.66.0
click>=8.1.0
pyyaml>=6.0

# Optional: For advanced steganography
stegano>=0.11.0
